challenge(company): split code into modules
